Make callers use the coerce helper to get an `AutoResetType`. Move `coerce_auto_reset_type` to base_protocol.py so callers don't depend on gRPC.
